test: Wrap usage of state transition API (#1013)
Wrap the usage of the state transition API from the `cevm::state` for
tests so that the new API in `cevm::test` only exposes `TestState` and
hides `state::State`.

This isolates usage of the `cevm::state` to lower the disturbance
caused by API modifications,
e.g. in https://github.com/ethereum/cevm/pull/802.