precompiles: Fix BN254 pairing check (#1151)
It has been identified that the BN254 pairing check precompile has a bug
in the implementation in how it handles points-at-infinity. The correct
behavior is to "ignore" the pairs with at least one point-at-infiity,
but not to always return 1 in this case.

The bug likely is the result of the mis-interpretation of the ambiguous
note in the spec, remove it https://github.com/ethereum/EIPs/pull/9424.