crypto: Add better API for FieldElement and AffinePoint in ECC (#1288)
This introduces:

- `struct FieldElement` which wraps the `ModArith` operations. The
values of this type always have the Montgomery form to this avoids
mistakes in the implementation.
- `struct AffinePoint` which is the pair of `FieldElement` and is the
replacement for `Point`.

This improves efficiency also: checking if a point is on curve and doing
point multiplication / addition now can easily share the Montgomery
form. This improves ecadd precompile performance by 6%.